+++ tests/unit/clients/test_json_rpc_client.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 
 from unittest import TestCase
 
-from xrpl.asyncio.clients.exceptions import XRPLRequestFailureException
 from xrpl.clients import JsonRpcClient
 from xrpl.models.requests import ServerInfo
 
@@ -16,10 +15,3 @@ class TestJsonRpcClient(TestCase):
         JSON_RPC_URL = "https://s.altnet.rippletest.net:51234/"
         client = JsonRpcClient(JSON_RPC_URL)
         client.request(ServerInfo())
-
-    def test_json_rpc_client_invalid_url(self: TestJsonRpcClient) -> None:
-        # Invalid URL
-        JSON_RPC_URL = "https://s2.ripple.com:51233/"
-        with self.assertRaises(XRPLRequestFailureException):
-            client = JsonRpcClient(JSON_RPC_URL)
-            client.request(ServerInfo())